+++ audio_conversion/src/AudioReformatter.cpp
@@ -67,8 +67,8 @@ status_t AudioReformatter::configure(const SampleSpec &ssSrc, const SampleSpec &
 
 status_t AudioReformatter::convertS16toS24over32(const void *src,
                                                  void *dst,
-                                                 const uint32_t inFrames,
-                                                 uint32_t *outFrames)
+                                                 const size_t inFrames,
+                                                 size_t *outFrames)
 {
     uint32_t frameIndex;
     const int16_t *src16 = (const int16_t *)src;
@@ -89,8 +89,8 @@ status_t AudioReformatter::convertS16toS24over32(const void *src,
 
 status_t AudioReformatter::convertS24over32toS16(const void *src,
                                                  void *dst,
-                                                 const uint32_t inFrames,
-                                                 uint32_t *outFrames)
+                                                 const size_t inFrames,
+                                                 size_t *outFrames)
 {
     const uint32_t *src32 = (const uint32_t *)src;
     int16_t *dst16 = (int16_t *)dst;
