+++ audio_conversion/src/AudioReformatter.hpp
@@ -61,8 +61,8 @@ private:
      */
     android::status_t convertS16toS24over32(const void *src,
                                             void *dst,
-                                            const uint32_t inFrames,
-                                            uint32_t *outFrames);
+                                            const size_t inFrames,
+                                            size_t *outFrames);
 
     /**
      * Converts (Reformats) audio samples.
@@ -78,8 +78,8 @@ private:
      */
     android::status_t convertS24over32toS16(const void *src,
                                             void *dst,
-                                            const uint32_t inFrames,
-                                            uint32_t *outFrames);
+                                            const size_t inFrames,
+                                            size_t *outFrames);
 
     /**
      * Used to do 8-bits right shitfs during reformatting operation.
