+++ audio_conversion/src/AudioResampler.hpp
@@ -79,8 +79,8 @@ private:
      */
     virtual android::status_t convert(const void *src,
                                       void **dst,
-                                      uint32_t inFrames,
-                                      uint32_t *outFrames);
+                                      size_t inFrames,
+                                      size_t *outFrames);
 
     /**
      * Resampler to use for all conversions.
